Fix spaces in stamped values, refactor internal scripts to avoid code duplication
