+++ internal/runWithVersionInfo.mjs
@@ -1,6 +1,6 @@
 import { spawn } from 'node:child_process'
-import fs from 'node:fs/promises'
 import process from 'node:process'
+import { readStableStatusFile } from './stableStatus.mjs'
 
 if (process.env.JS_BINARY__EXECROOT) {
   process.chdir(process.env.JS_BINARY__EXECROOT)
@@ -10,26 +10,6 @@ function usage() {
   throw new Error(
     `Invalid arguments.\nusage:\n  ${process.argv[0]} ${process.argv[1]} [--status-file <stable-status.txt>] <executable> <args...>`,
   )
-}
-
-/**
- * Parse Bazel stable-status.txt into a key/value map.
- * Keys are returned without the leading 'STABLE_' prefix.
- *
- * @param {string} content
- * @returns {Record<string, string>}
- */
-function parseStableStatus(content) {
-  /** @type {Record<string, string>} */
-  const vars = {}
-  for (const line of content.split(/\r?\n/)) {
-    const [key, value] = line.split(' ', 2)
-    if (key && key.startsWith('STABLE_ENSO')) {
-      const envName = key.slice('STABLE_'.length)
-      vars[envName] = value
-    }
-  }
-  return vars
 }
 
 function spawnExecutable(executable, args, env) {
@@ -63,8 +43,7 @@ const executable = process.argv[idx]
 const execArgs = process.argv.slice(idx + 1)
 
 if (statusFilePath != null) {
-  const content = await fs.readFile(statusFilePath, 'utf8')
-  const stableVars = parseStableStatus(content)
+  const stableVars = await readStableStatusFile(statusFilePath)
 
   // TODO: sbt should really use ENSO_IDE_VERSION and ENSO_IDE_EDITION variables.
   // Or, IDE build should use more universal ENSO_VERSION and ENSO_EDITION
